requirements.txt
requests
raven
mailchimp3
python-dotenv